Rewrite `protocol` crate
The `protocol`  crate was entirely rewritten to be both more ergonomic
and efficient. A proc-macro derive has been added to auto-derive what
used to be the `Protocol` trait and which is now split in two traits
`Encode` and `Decode`.

The protocol generation tool has been rewritten in Rust as well (the
code is still *not great*, though).