Re-implement `Framed` from tokio-codec to fit our needs
We can now encode Dofus messages directly onto the very buffer that
will be written to the tcp stream. For this to be possible,
I needed to change `Encode::encode` from taking a `&mut Vec<u8>` to
taking a `&mut bytes::BytesMut`. Btw, I updated the `bytes`
dependency to point to the master branch of the crate repository, as
a lot of useful changes are happening there.